+++ tests/viadot_tests/unit/test_hubspot.py
@@ -1,203 +1,225 @@
+import json
 import unittest
-from unittest import mock
+from datetime import datetime
+from io import StringIO
+from unittest.mock import MagicMock, patch
 
 import pandas as pd
 import pytest
+from requests.models import Response
 
-from viadot.exceptions import APIError
+from viadot.exceptions import APIError, CredentialError
 from viadot.sources import Hubspot
 from viadot.sources.hubspot import HubspotCredentials
 
-
-@pytest.fixture(scope="function")
-def var_dictionary():
-    variables = {
-        "credentials": {
-            "token": "xxxxxxxxxxxxxxxxxxxxxxx",
-        },
-        "filters": [
-            {
-                "filters": [
-                    {
-                        "propertyName": "createdate",
-                        "operator": "BETWEEN",
-                        "highValue": "1642636800000",
-                        "value": "1641995200000",
-                        "pruebas": "2024-06-24",
-                    },
-                    {
-                        "propertyName": "email",
-                        "operator": "CONTAINS_TOKEN",
-                        "value": "*@xxxxxxx.xx",
-                    },
-                ]
-            }
-        ],
-    }
-    return variables
-
-
-class MockClass:
-    status_code = 200
-
-    def json():
-        test = {
-            "objects": [
+variables = {
+    "credentials": {"token": "fake_token"},
+    "filters": [
+        {
+            "filters": [
                 {
-                    "id": 9999999999,
-                    "createdAt": 1719201702051,
-                    "updatedAt": 1719201702051,
-                    "publishedAt": None,
-                    "path": None,
-                    "name": None,
-                    "values": {"1": "xxxxxx", "2": "xxxxx"},
-                    "isSoftEditable": False,
-                    "childTableId": 0,
+                    "propertyName": "createdate",
+                    "operator": "GTE",
+                    "value": "2021-01-01",
                 }
-            ],
-            "total": 164,
-            "limit": 1000,
-            "offset": 0,
-            "message": None,
-            "totalCount": 164,
+            ]
         }
-        return test
-
-
-class MockClassF:
-    status_code = 200
-
-    def json():
-        test = {
-            "results": {
-                "objects": [
-                    {
-                        "id": 9999999999,
-                        "createdAt": 1719201702051,
-                        "updatedAt": 1719201702051,
-                        "publishedAt": None,
-                        "path": None,
-                        "name": None,
-                        "values": {"1": "xxxxxx", "2": "xxxxx"},
-                        "isSoftEditable": False,
-                        "childTableId": 0,
-                    }
-                ],
-                "total": 164,
-                "limit": 1000,
-                "offset": 0,
-                "message": None,
-                "totalCount": 164,
-            }
-        }
-        return test
-
-
-class MockClassException:
-    status_code = 404
-    content = b""
-
-    def json():
-        return None
-
-
-@mock.patch("viadot.sources.hubspot.handle_api_response", return_value=MockClass)
-@pytest.mark.connect
-def test_hubspot_api_response(mock_connection, var_dictionary):
-    hp = Hubspot(credentials=var_dictionary["credentials"])
-    hp.api_connection(
-        endpoint="hubdb/api/v2/tables/99999999",
-    )
-
-    mock_connection.assert_called_once()
-    assert isinstance(hp.full_dataset, list)
-
-
-@mock.patch("viadot.sources.hubspot.handle_api_response", return_value=MockClassF)
-@pytest.mark.connect
-def test_hubspot_api_response_filters(mock_connection, var_dictionary):
-    hp = Hubspot(credentials=var_dictionary["credentials"])
-    hp.api_connection(
-        endpoint="hubdb/api/v2/tables/99999999",
-        filters=var_dictionary["filters"],
-    )
-
-    mock_connection.assert_called_once()
-    assert isinstance(hp.full_dataset, dict)
-
-
-@mock.patch(
-    "viadot.sources.hubspot.handle_api_response", return_value=MockClassException
-)
-@pytest.mark.connect
-def test_hubspot_api_error(mock_connection, var_dictionary):
-    hp = Hubspot(credentials=var_dictionary["credentials"])
-    with pytest.raises(APIError):
-        hp.api_connection(
-            endpoint="hubdb/api/v2/tables/99999999",
-        )
-
-    mock_connection.assert_called_once()
-
-
-@mock.patch("viadot.sources.hubspot.handle_api_response", return_value=MockClass)
-@pytest.mark.response
-def test_hubspot_api_df_response(mock_connection, var_dictionary):
-    hp = Hubspot(credentials=var_dictionary["credentials"])
-    hp.api_connection(
-        endpoint="hubdb/api/v2/tables/99999999",
-    )
-    df = hp.to_df()
-
-    viadot_set = {"_viadot_source", "_viadot_downloaded_at_utc"}
-
-    mock_connection.assert_called_once()
-    assert set(df.columns).issuperset(viadot_set)
-    assert isinstance(df, pd.DataFrame)
+    ],
+}
 
 
 class TestHubspotCredentials:
-    @pytest.fixture(scope="function")
-    def hubspotcredentials(self):
-        return HubspotCredentials()
+    """Test Hubspot Credentials Class."""
+
+    @pytest.mark.basic
+    def test_hubspot_credentials(self):
+        """Test Hubspot credentials."""
+        HubspotCredentials(token="test_token")
 
 
-class TestHubspot:
-    @pytest.fixture(scope="function")
-    def hubspot(self):
-        return Hubspot()
+class TestHubspot(unittest.TestCase):
+    """Test Hubspot Class."""
 
-    def test__date_to_unixtimestamp(self, hubspot):
-        # TODO [use mock.patch, assert]: Implement test for Hubspot._date_to_unixtimestamp (line 121)
-        pass
+    @classmethod
+    def setUpClass(cls):
+        cls.hubspot_instance = Hubspot(credentials=variables["credentials"])
 
-    def test__get_api_url(self, hubspot):
-        # TODO [use mock.patch, assert]: Implement test for Hubspot._get_api_url (line 139)
-        pass
+    @patch("viadot.sources.hubspot.get_source_credentials", return_value=None)
+    @pytest.mark.basic
+    def test_init_no_credentials(self, mock_get_source_credentials):
+        """Test raise error without credentials."""
+        with pytest.raises(CredentialError):
+            Hubspot()
 
-    def test__format_filters(self, hubspot):
-        # TODO [use mock.patch, assert]: Implement test for Hubspot._format_filters (line 175)
-        pass
+        mock_get_source_credentials.assert_called_once()
 
-    def test__get_api_body(self, hubspot):
-        # TODO [use mock.patch, assert]: Implement test for Hubspot._get_api_body (line 201)
-        pass
+    @pytest.mark.functions
+    def test_date_to_unixtimestamp(self):
+        """Test Hubspot `_date_to_unixtimestamp` function."""
+        date_str = "2021-01-01"
+        expected_timestamp = int(
+            datetime.strptime(date_str, "%Y-%m-%d").timestamp() * 1000
+        )
+        result = self.hubspot_instance._date_to_unixtimestamp(date_str)
+        self.assertEqual(result, expected_timestamp)
 
-    def test__api_call(self, hubspot):
-        # TODO [use mock.patch, assert]: Implement test for Hubspot._api_call (line 248)
-        pass
+    @pytest.mark.functions
+    def test_get_api_url(self):
+        """Test Hubspot `_get_api_url` function."""
+        endpoint = "deals"
+        filters = None
+        properties = ["property1", "property2"]
+        expected_url = (
+            f"https://api.hubapi.com/crm/v3/objects/{endpoint}/"
+            + "?limit=100&properties=property1,property2&"
+        )
+        result = self.hubspot_instance._get_api_url(
+            endpoint=endpoint, filters=filters, properties=properties
+        )
+        self.assertEqual(result, expected_url)
 
-    def test__get_offset_from_response(self, hubspot):
-        # TODO [use mock.patch, assert]: Implement test for Hubspot._get_offset_from_response (line 287)
-        pass
+    @pytest.mark.functions
+    def test_format_filters(self):
+        """Test Hubspot `_format_filters` function."""
+        filters = variables["filters"]
+        formatted_filters = self.hubspot_instance._format_filters(filters)
+        self.assertIsInstance(formatted_filters, list)
 
-    def test_api_connection(self, hubspot):
-        # TODO [use mock.patch, assert]: Implement test for Hubspot.api_connection (line 313)
-        pass
+    @pytest.mark.functions
+    def test_get_api_body(self):
+        """Test Hubspot `_get_api_body` function."""
+        filters = variables["filters"]
+        expected_body = json.dumps({"filterGroups": filters, "limit": 100})
+        result = self.hubspot_instance._get_api_body(filters)
 
-    def test_to_df(self, hubspot):
-        # TODO [use mock.patch, assert]: Implement test for Hubspot.to_df (line 395)
-        pass
+        self.assertEqual(result, expected_body)
+
+    @pytest.mark.connect
+    @patch("viadot.sources.hubspot.handle_api_response")
+    def test_api_call_success(self, mock_handle_api_response):
+        """Test Hubspot `_api_call` method."""
+        mock_response = MagicMock(spec=Response)
+        mock_response.status_code = 200
+        mock_response.json.return_value = {"results": [{"id": "123"}]}
+        mock_handle_api_response.return_value = mock_response
+
+        url = "https://api.hubapi.com/crm/v3/objects/deals/?limit=100&"
+        result = self.hubspot_instance._api_call(url=url, method="GET")
+
+        self.assertEqual(result, {"results": [{"id": "123"}]})
+        mock_handle_api_response.assert_called_once()
+
+    @pytest.mark.connect
+    @patch("viadot.sources.hubspot.handle_api_response")
+    def test_api_call_error(self, mock_handle_api_response):
+        """Test Hubspot `_api_call` method failure."""
+        mock_response = MagicMock(spec=Response)
+        mock_response.status_code = 500
+        mock_response.content = b"Internal Server Error"
+        mock_handle_api_response.return_value = mock_response
+
+        url = "https://api.hubapi.com/crm/v3/objects/deals/?limit=100&"
+
+        with self.assertRaises(APIError):
+            self.hubspot_instance._api_call(url=url, method="GET")
+
+    @pytest.mark.functions
+    def test_get_offset_from_response(self):
+        """Test Hubspot `_get_offset_from_response` function."""
+        response_with_paging = {"paging": {"next": {"after": "123"}}}
+        response_with_offset = {"offset": "456"}
+
+        offset_type, offset_value = self.hubspot_instance._get_offset_from_response(
+            response_with_paging
+        )
+        self.assertEqual((offset_type, offset_value), ("after", "123"))
+
+        offset_type, offset_value = self.hubspot_instance._get_offset_from_response(
+            response_with_offset
+        )
+        self.assertEqual((offset_type, offset_value), ("offset", "456"))
+
+        offset_type, offset_value = self.hubspot_instance._get_offset_from_response({})
+        self.assertEqual((offset_type, offset_value), (None, None))
+
+    @pytest.mark.connect
+    @patch("viadot.sources.hubspot.handle_api_response")
+    def test_api_connection_with_filters(self, mock_handle_api_response):
+        """Test Hubspot `api_connection` method, with filters."""
+        mock_response = MagicMock(spec=Response)
+        mock_response.status_code = 200
+        mock_response.json.return_value = {"results": [{"id": "123"}]}
+        mock_handle_api_response.return_value = mock_response
+
+        endpoint = "deals"
+        filters = variables["filters"]
+        properties = ["property1"]
+        self.hubspot_instance.api_connection(
+            endpoint=endpoint, filters=filters, properties=properties
+        )
+
+        self.assertIsNotNone(self.hubspot_instance.full_dataset)
+        self.assertGreater(len(self.hubspot_instance.full_dataset), 0)
+
+    @pytest.mark.connect
+    @patch("viadot.sources.hubspot.handle_api_response")
+    def test_api_connection_without_filters(self, mock_handle_api_response):
+        """Test Hubspot `api_connection` method, without filters."""
+        mock_response = MagicMock(spec=Response)
+        mock_response.status_code = 200
+        mock_response.json.return_value = {"results": [{"id": "123"}]}
+        mock_handle_api_response.return_value = mock_response
+
+        endpoint = "deals"
+        filters = None
+        properties = ["property1"]
+        self.hubspot_instance.api_connection(
+            endpoint=endpoint, filters=filters, properties=properties
+        )
+
+        self.assertIsNotNone(self.hubspot_instance.full_dataset)
+        self.assertGreater(len(self.hubspot_instance.full_dataset), 0)
+
+    @pytest.mark.functions
+    @patch("viadot.sources.hubspot.super")
+    def test_to_df(self, mock_super):
+        """Test Hubspot `to_df` function."""
+        mock_super().to_df = MagicMock()
+        self.hubspot_instance.full_dataset = [{"id": "123"}]
+        result_df = self.hubspot_instance.to_df()
+        result_df.drop(
+            columns=["_viadot_source", "_viadot_downloaded_at_utc"],
+            inplace=True,
+            axis=1,
+        )
+
+        expected_df = pd.DataFrame([{"id": "123"}])
+        pd.testing.assert_frame_equal(result_df, expected_df)
+        mock_super().to_df.assert_called_once()
+
+    @pytest.mark.functions
+    @patch("viadot.sources.hubspot.pd.read_json")
+    @patch("viadot.sources.hubspot.super")
+    def test_to_df_empty(self, mock_super, mock_read_json):
+        """Test Genesys `to_df` method, checking emptiness."""
+        mock_super().to_df = MagicMock()
+        mock_read_json.return_value = pd.DataFrame()
+        self.hubspot_instance.full_dataset = StringIO("{}")
+
+        with patch.object(
+            self.hubspot_instance, "_handle_if_empty"
+        ) as mock_handle_if_empty:
+            result_df = self.hubspot_instance.to_df()
+            result_df.drop(
+                columns=["_viadot_source", "_viadot_downloaded_at_utc"],
+                inplace=True,
+                axis=1,
+            )
+            mock_handle_if_empty.assert_called_once_with(
+                if_empty="warn", message="The response does not contain any data."
+            )
+            self.assertTrue(result_df.empty)
+            mock_super().to_df.assert_called_once()
 
 
 if __name__ == "__main__":
